hid-replay: Add 'accurate-timing' option
At the moment, hid-replay will not bother sleeping if two events are
seperated by less than 500 microseconds, and will also truncate event
intervals that exceed 3 seconds to be 3 seconds. This is generally fine,
but if accurate event timing is important, there should be a way to
disable this behavior. The 'accurate-timing' option fills this need by
always sleeping for the required amount of time (unless it is zero or
in the past).

Signed-off-by: Jason Gerecke <jason.gerecke@wacom.com>