hid-replay: Remove timing drift when using accurate timing
The '--accurate-timing' option can exibit slow-but-sure timing drift as
small sleep time differences add up. To mitigate this, keep an absolute
reference time that lets us compare the amount of time we've actually
spent replaying the recording vs the recording's total elapsed time.
To have the non-accurate timing mode continue to work, we modify the
reference time to make it appear that we've been replaying for longer
than we actually have.

Signed-off-by: Jason Gerecke <jason.gerecke@wacom.com>